Change width and height types in QrCode struct

+++ src/lib.rs
@@ -83,8 +83,8 @@ pub struct QrCode {
     content: Vec<Color>,
     version: Version,
     ec_level: EcLevel,
-    width: usize,
-    height: usize,
+    width: u8,
+    height: u8,
 }
 
 impl QrCode {
@@ -361,7 +361,7 @@ impl QrCode {
     /// assert_eq!(code.width(), 27);
     /// ```
     #[must_use]
-    pub const fn width(&self) -> usize {
+    pub const fn width(&self) -> u8 {
         self.width
     }
 
@@ -378,7 +378,7 @@ impl QrCode {
     /// assert_eq!(code.height(), 13);
     /// ```
     #[must_use]
-    pub const fn height(&self) -> usize {
+    pub const fn height(&self) -> u8 {
         self.height
     }
